+++ tomcat/src/main/java/org/apache/coyote/domain/HttpRequest.java
@@ -1,4 +1,4 @@
-package org.apache.coyote;
+package org.apache.coyote.domain;
 
 import java.util.HashMap;
 import java.util.Map;
