+++ tomcat/src/main/java/org/apache/coyote/domain/MyHttpResponse.java
@@ -1,4 +1,4 @@
-package org.apache.coyote;
+package org.apache.coyote.domain;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
