+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -9,9 +9,9 @@ import java.util.Optional;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.UncheckedServletException;
 import nextstep.jwp.model.User;
-import org.apache.coyote.FilePath;
-import org.apache.coyote.HttpRequest;
-import org.apache.coyote.MyHttpResponse;
+import org.apache.coyote.domain.FilePath;
+import org.apache.coyote.domain.HttpRequest;
+import org.apache.coyote.domain.MyHttpResponse;
 import org.apache.coyote.Processor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
